Decode github deploy private key
Note that the password used to decode is stored in the shmig
.travis.yml, since that is the project that is ultimately being
built by travis.

This shmig_test is checked out as part of that build.

Since the password is encrypted with the private key that only
Travis knows, and that key varies by repository, we put the key
in the Travis .tavis.yml.